Refactor tasks into components
