+++ website/src/pages/create/initial_prompt.tsx
@@ -1,10 +1,9 @@
 import { Container } from "@chakra-ui/react";
 import { useColorMode } from "@chakra-ui/react";
+import Head from "next/head";
 import { useEffect, useState } from "react";
 import { LoadingScreen } from "src/components/Loading/LoadingScreen";
-import { TaskControls } from "src/components/Survey/TaskControls";
-import { TrackedTextarea } from "src/components/Survey/TrackedTextarea";
-import { TwoColumnsWithCards } from "src/components/Survey/TwoColumnsWithCards";
+import { Task } from "src/components/Tasks/Task";
 import fetcher from "src/lib/fetcher";
 import poster from "src/lib/poster";
 import useSWRImmutable from "swr/immutable";
@@ -12,7 +11,6 @@ import useSWRMutation from "swr/mutation";
 
 const InitialPrompt = () => {
   const [tasks, setTasks] = useState([]);
-  const [inputText, setInputText] = useState("");
 
   const { isLoading, mutate } = useSWRImmutable("/api/new_task/initial_prompt ", fetcher, {
     onSuccess: (data) => {
@@ -33,26 +31,6 @@ const InitialPrompt = () => {
     }
   }, [tasks]);
 
-  const submitResponse = (task: { id: string }) => {
-    const text = inputText.trim();
-    trigger({
-      id: task.id,
-      update_type: "text_reply_to_message",
-      content: {
-        text,
-      },
-    });
-  };
-
-  const fetchNextTask = () => {
-    setInputText("");
-    mutate();
-  };
-
-  const textChangeHandler = (event: React.ChangeEvent<HTMLTextAreaElement>) => {
-    setInputText(event.target.value);
-  };
-
   const { colorMode } = useColorMode();
   const mainBgClasses = colorMode === "light" ? "bg-slate-300 text-gray-800" : "bg-slate-900 text-white";
 
@@ -65,25 +43,13 @@ const InitialPrompt = () => {
   }
 
   return (
-    <div className={`p-12 ${mainBgClasses}`}>
-      <TwoColumnsWithCards>
-        <>
-          <h5 className="text-lg font-semibold">Start a conversation</h5>
-          <p className="text-lg py-1">Create an initial message to send to the assistant</p>
-        </>
-        <>
-          <h5 className="text-lg font-semibold">Provide the initial prompt</h5>
-          <TrackedTextarea
-            text={inputText}
-            onTextChange={textChangeHandler}
-            thresholds={{ low: 20, medium: 40, goal: 50 }}
-            textareaProps={{ placeholder: "Question, task, greeting or similar..." }}
-          />
-        </>
-      </TwoColumnsWithCards>
-
-      <TaskControls tasks={tasks} onSubmitResponse={submitResponse} onSkip={fetchNextTask} />
-    </div>
+    <>
+      <Head>
+        <title>Reply as Assistant</title>
+        <meta name="description" content="Reply as Assistant." />
+      </Head>
+      <Task tasks={tasks} trigger={trigger} mutate={mutate} mainBgClasses={mainBgClasses} />
+    </>
   );
 };
